+++ common/src/main/java/net/mehvahdjukaar/every_compat/EveryCompat.java
@@ -244,9 +244,9 @@ public abstract class EveryCompat {
     }
 
     public synchronized static void addOptionalModule(String modId, Supplier<Class<? extends CompatModule>> moduleClass) {
-        if (PlatHelper.isModLoaded(modId)
-                && !MODULES_BLACKLIST.get().contains(modId)
-                && !ENTRY_SETS_BLACKLIST.get().contains(modId + ":.*")) {
+        boolean isModuleNotBlacklisted = !(MODULES_BLACKLIST.get().contains(modId) || ENTRY_SETS_BLACKLIST.get().contains(modId + ":.*"));
+
+        if (PlatHelper.isModLoaded(modId) && isModuleNotBlacklisted) {
             try {
                 Class<? extends CompatModule> klazz = moduleClass.get();
 
@@ -269,6 +269,17 @@ public abstract class EveryCompat {
         }
     }
 
+    @SafeVarargs
+    public static void addMultipleOptional(String modId, Supplier<Class<? extends CompatModule>>... klazzes) {
+        boolean isModuleNotBlacklisted = !(MODULES_BLACKLIST.get().contains(modId) || ENTRY_SETS_BLACKLIST.get().contains(modId + ":.*"));
+
+        if (PlatHelper.isModLoaded(modId) && isModuleNotBlacklisted) {
+            for (var klazz : klazzes) {
+                addOptionalModule(modId, klazz);
+            }
+        }
+    }
+
     private static void addError(CompatModule module, Throwable t) {
         if (module == null) {
             EveryCompat.LOGGER.error("Tried to log an error for a null module", t);
@@ -287,13 +298,6 @@ public abstract class EveryCompat {
         }
         ERRORED.put(Preconditions.checkNotNull(module, "Module cannot be null"),
                 Preconditions.checkNotNull(t, "Throwable cannot be null"));
-    }
-
-    @SafeVarargs
-    public static void addMultipleOptional(String modId, Supplier<Class<? extends CompatModule>>... klazzes) {
-        for (var klazz : klazzes) {
-            addOptionalModule(modId, klazz);
-        }
     }
 
     private static @NotNull CompatModule instantiateModuleClass(String modId, Class<? extends CompatModule> klazz)
@@ -347,7 +351,9 @@ public abstract class EveryCompat {
                     event.add(e.getKey(), ee.toArray(ItemLike[]::new));
                 }
             }
-        } else {
+        }
+
+        if (!ECConfigs.NO_MOD_CREATIVE_TAB.get()) {
             forAllModules(m -> m.registerItemsToExistingTabs(event));
         }
     }
@@ -362,7 +368,11 @@ public abstract class EveryCompat {
                         },
                         entry -> {
                             var message = entry.getValue().getMessage();
-                            return message != null ? message : "Failed to get error message");
+                            var cause = entry.getValue().getCause();
+
+                            if (message != null) return message;
+                            else if (cause != null) return cause.toString();
+                            else return "Failed to get error message";
                         }
                 ));
     }