+++ common/src/main/java/net/mehvahdjukaar/every_compat/misc/BlockTypeCycleItemRenderer.java
@@ -87,7 +87,7 @@ public abstract class BlockTypeCycleItemRenderer<T extends BlockType> extends It
         int tm = time % (size+1);
         if (tm != lastTime) {
 
-            ItemLike v = null;
+            ItemLike itemLike = null;
             do {
                 var l = (this.lastIndex + 1) % size;
                 // this.woodIndex = (this.woodIndex + 1);
@@ -96,13 +96,14 @@ public abstract class BlockTypeCycleItemRenderer<T extends BlockType> extends It
                 String key = childKeys.get(lastIndex);
                 var vv = moddedTypes.get(typeIndex % moddedTypes.size()).getChild(key);
                 if (vv instanceof ItemLike il) {
-                    v = il;
+                    itemLike = il;
                 }
-            } while (v == null);
+            } while (itemLike == null);
 
-            this.currentStack = v.asItem().getDefaultInstance();
+            this.currentStack = itemLike.asItem().getDefaultInstance();
         }
         this.lastTime = tm;
         return currentStack;
     }
+
 }