+++ common/src/main/java/net/mehvahdjukaar/every_compat/misc/WoodTypeCycleItemRenderer.java
@@ -6,4 +6,5 @@ public class WoodTypeCycleItemRenderer extends BlockTypeCycleItemRenderer<WoodTy
     public WoodTypeCycleItemRenderer() {
         super(WoodType.class);
     }
+
 }